Handle `State` Mixin deprecation/removal
Troposphere has heavily used `Router.State` to make available params of
a URL (projectId, imageId, etc). This is no longer available [0, 1].

For components that *directly* participate as a "route component", calls
to `getPath()` or `getParams()` are replaced by `this.props.path` and
`this.props.params`.

What's a "route component?"

If a component is mentioned in a route definition as the component
attribute - it is a "route component"

```
<Router>
  <Route component={Master} >
    <Route path="images" component={ImagesMaster}>
      <IndexRoute component={ImageListPage} />
      <Route path="search" component={ImageListPage}/>
      <Route path="tags" component={ImageTagsPage} />
      <Route path=":imageId" component={ImageDetailsPage} />
    </Route>
    <Route path="help" component={HelpPage} />
    <IndexRedirect to="images" />
    <Route path="*" component={NotFoundPage} />
  </Route>
</Router>
```

Route components: Master, ImagesMaster, ImageListPage, ImageTagsPage,
 ImageDetailsPage, HelpPage, NotFoundPage

*IMPORTANT*

If a component is **not** a route component, then the approach to
provide `params` is to just `this.context`.

This is done by specifying `contextTypes`. An example could be found
within
`components/projects/details/resources/tableData/instance/Name.jsx`

```
export default React.createClass({
     displayName: "Name",

     contextTypes: {
         params: React.PropTypes.object
     },
//....
});
```

tread lightly kind soul ...

[0] https://github.com/ReactTraining/react-router/blob/master/upgrade-guides/v1.0.0.md#state-mixin
[1] https://github.com/ReactTraining/react-router/blob/master/upgrade-guides/v2.0.0.md#mixins-are-deprecated