Ensure ImageListPage gets data changes
During the upgrade, it became clear that this page was not receiving
state updates when TagStore and HelpLinkStore were done fetching. It is
not clear why this was previously not encountered, but we have seen in
other scenarios where the data was eager loaded by other components &
may have made it more difficult to see obvious "fetch" issues like
this. It seems that using the "updateState" pattern adopted within
other page components makes sense - but should be evaluated during
pull request review.