Add instance action "User Customizations" to allow additional playbook runs
This is a new feature that allows users to deploy additional playbooks on a running instance.
Currently, it reads from a JSON file and presents the user with a checklist of palybooks to run.
When submitted, the list of selected playbooks will be sent to Atmosphere and deployed to the instance.