Show previously selected customizations in modal and disable checkbox
To make it possible to add UserCustomizations to the instance during launch, I had to make it an attribute to Atmosphere's instance object.
This also allowed me to access the previously installed UserCustomizations so I can show them in the modal. The checkboxes are disabled so the user doesn't expect things to be uninstalled if the deselect a box.